requirements.txt
pandas==2.2.3
openai==1.78.1
python-dotenv==1.1.0
langchain==0.2.1
tiktoken==0.5.2
langchain_community
tiktoken
pyautogen
